Fix a broken test by refactoring the division_bloc.dart
